# Pool blake3.New() in digest.Compute