blake3.New() allocates over 10KiB. Pooling these makes digest.Compute much faster for blake3 resources.

Benchmark results:
```
                               │     base      │               pooled                │
                               │    sec/op     │    sec/op     vs base               │
DigestCompute/SHA256/1-24         193.4n ± 10%   190.6n ± 11%        ~ (p=0.644 n=7)
DigestCompute/BLAKE3/1-24        1306.0n ± 18%   204.3n ± 10%  -84.36% (p=0.001 n=7)
DigestCompute/SHA256/10-24        184.3n ± 11%   191.2n ± 36%        ~ (p=0.209 n=7)
DigestCompute/BLAKE3/10-24       1240.0n ± 24%   206.8n ± 13%  -83.32% (p=0.001 n=7)
DigestCompute/SHA256/100-24       210.7n ± 22%   237.7n ± 17%        ~ (p=0.165 n=7)
DigestCompute/BLAKE3/100-24      1512.0n ± 28%   258.4n ±  1%  -82.91% (p=0.001 n=7)
DigestCompute/SHA256/1000-24      560.3n ±  5%   569.6n ±  3%        ~ (p=0.805 n=7)
DigestCompute/BLAKE3/1000-24     1883.0n ± 45%   916.6n ±  1%  -51.32% (p=0.001 n=7)
DigestCompute/SHA256/10000-24     3.909µ ±  1%   3.948µ ±  1%   +1.00% (p=0.048 n=7)
DigestCompute/BLAKE3/10000-24     4.446µ ±  4%   3.083µ ±  8%  -30.66% (p=0.001 n=7)
DigestCompute/SHA256/100000-24    37.59µ ±  1%   37.58µ ±  1%        ~ (p=0.805 n=7)
DigestCompute/BLAKE3/100000-24    19.49µ ±  5%   17.39µ ± 20%  -10.80% (p=0.026 n=7)
geomean                           1.583µ         925.1n        -41.55%

                               │     base     │               pooled                │
                               │     B/op     │    B/op     vs base                 │
DigestCompute/SHA256/1-24          400.0 ± 0%   400.0 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/1-24        11184.0 ± 0%   304.0 ± 0%  -97.28% (p=0.001 n=7)
DigestCompute/SHA256/10-24         400.0 ± 0%   400.0 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/10-24       11184.0 ± 0%   304.0 ± 0%  -97.28% (p=0.001 n=7)
DigestCompute/SHA256/100-24        400.0 ± 0%   400.0 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/100-24      11184.0 ± 0%   304.0 ± 0%  -97.28% (p=0.001 n=7)
DigestCompute/SHA256/1000-24       400.0 ± 0%   400.0 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/1000-24     11184.0 ± 0%   304.0 ± 0%  -97.28% (p=0.001 n=7)
DigestCompute/SHA256/10000-24      400.0 ± 0%   400.0 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/10000-24    11184.0 ± 0%   304.0 ± 0%  -97.28% (p=0.001 n=7)
DigestCompute/SHA256/100000-24     400.0 ± 0%   400.0 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/100000-24   11184.0 ± 0%   304.0 ± 0%  -97.28% (p=0.001 n=7)
geomean                          2.066Ki        348.7       -83.51%
¹ all samples are equal

                               │    base    │               pooled                │
                               │ allocs/op  │ allocs/op   vs base                 │
DigestCompute/SHA256/1-24        6.000 ± 0%   6.000 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/1-24        7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.001 n=7)
DigestCompute/SHA256/10-24       6.000 ± 0%   6.000 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/10-24       7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.001 n=7)
DigestCompute/SHA256/100-24      6.000 ± 0%   6.000 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/100-24      7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.001 n=7)
DigestCompute/SHA256/1000-24     6.000 ± 0%   6.000 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/1000-24     7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.001 n=7)
DigestCompute/SHA256/10000-24    6.000 ± 0%   6.000 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/10000-24    7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.001 n=7)
DigestCompute/SHA256/100000-24   6.000 ± 0%   6.000 ± 0%        ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/100000-24   7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.001 n=7)
geomean                          6.481        6.000        -7.42%
¹ all samples are equal
```